scripts/requirements.txt
openpyxl>=3.1.0
azure-data-tables>=12.4.0
